Give the IbisDocApp an own class
- Avoids clutter
- Avoids too many lines in one file
